+++ apps/web/src/components/Widget/MembersWidget/index.js
@@ -1,3 +1,0 @@
-import MembersWidget from './MembersWidget'
-
-export default MembersWidget