+++ apps/web/src/components/Widget/TopicsWidget/index.js
@@ -1,3 +1,0 @@
-import TopicsWidget from './TopicsWidget'
-
-export default TopicsWidget